net, l2_bridge: Move secondary_network_vm
secondary_network_vm is generic enough to be reused.
Moving it to libl2bridge makes it available to all l2_bridge
sub-packages.
