+++ tests/network/l2_bridge/libl2bridge.py
@@ -3,18 +3,25 @@ import logging
 import re
 import time
 from ipaddress import ip_interface
+from typing import Final
 
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.resource import ResourceEditor
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
+from libs.net.cluster import ipv4_supported_cluster, ipv6_supported_cluster
 from libs.net.ip import random_ipv4_address
 from libs.net.vmspec import lookup_iface_status, lookup_iface_status_ip, wait_for_missing_iface_status
+from libs.vm.factory import base_vmspec, fedora_vm
+from libs.vm.spec import Affinity, CloudInitNoCloud, Interface, Multus, Network
+from libs.vm.vm import BaseVirtualMachine, add_volume_disk, cloudinitdisk_storage
+from tests.network.libs import cloudinit
 from tests.network.utils import update_cloud_init_extra_user_data
 from utilities import console
 from utilities.constants import (
     KUBEMACPOOL_MAC_CONTROLLER_MANAGER,
     LINUX_BRIDGE,
+    NODE_ROLE_KUBERNETES_IO,
     NODE_TYPE_WORKER_LABEL,
     SRIOV,
     TIMEOUT_1MIN,
@@ -359,6 +366,70 @@ def wait_for_no_packet_loss_after_connection(src_vm, dst_ip, interface=None):
     except TimeoutExpiredError:
         LOGGER.error(f"Ping from {src_vm.name} to {dst_ip} failed.")
         raise
+
+
+def secondary_network_vm(
+    namespace: str,
+    name: str,
+    client: DynamicClient,
+    nad_name: str,
+    secondary_iface_name: str,
+    secondary_iface_addresses: list[str],
+    affinity: Affinity | None = None,
+) -> BaseVirtualMachine:
+    """Create a Fedora VM with a masquerade primary interface and a secondary Linux bridge interface.
+
+    Args:
+        namespace: Namespace to deploy the VM in.
+        name: VM name.
+        client: Kubernetes dynamic client.
+        nad_name: NetworkAttachmentDefinition name for the secondary interface.
+        secondary_iface_name: Name of the secondary network interface in the VM spec.
+        secondary_iface_addresses: CIDR addresses to assign to the secondary interface via cloud-init.
+        affinity: Optional node or pod affinity rules for scheduling.
+    """
+    spec = base_vmspec()
+    spec.template.spec.domain.devices.interfaces = [  # type: ignore
+        Interface(name="default", masquerade={}),
+        Interface(name=secondary_iface_name, bridge={}),
+    ]
+    spec.template.spec.networks = [
+        Network(name="default", pod={}),
+        Network(name=secondary_iface_name, multus=Multus(networkName=nad_name)),
+    ]
+    if affinity:
+        spec.template.spec.affinity = affinity
+
+    ethernets = {}
+    primary = _masquerade_iface_cloud_init()
+    if primary:
+        ethernets["eth0"] = primary
+    ethernets["eth1"] = cloudinit.EthernetDevice(addresses=secondary_iface_addresses)
+
+    disk, volume = cloudinitdisk_storage(
+        data=CloudInitNoCloud(
+            networkData=cloudinit.asyaml(no_cloud=cloudinit.NetworkData(ethernets=ethernets)),
+            userData=cloudinit.format_cloud_config(userdata=cloudinit.UserData(users=[])),
+        )
+    )
+    spec.template.spec = add_volume_disk(vmi_spec=spec.template.spec, volume=volume, disk=disk)
+    return fedora_vm(namespace=namespace, name=name, client=client, spec=spec)
+
+
+def _masquerade_iface_cloud_init() -> cloudinit.EthernetDevice | None:
+    """Return cloud-init ethernet config for a masquerade (primary) interface.
+
+    Returns:
+        EthernetDevice with static IPv6 and optional DHCP4, or None if IPv6 is not supported.
+    """
+    if not ipv6_supported_cluster():
+        return None
+    return cloudinit.EthernetDevice(
+        addresses=["fd10:0:2::2/120"],
+        gateway6="fd10:0:2::1",
+        dhcp4=ipv4_supported_cluster(),
+        dhcp6=False,
+    )
 
 
 @contextlib.contextmanager
